tests/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
redis==5.0.1
pandas==2.1.3
numpy==1.26.2
python-telegram-bot==20.7 